http://adventofcode.com 2015 day 19 part 2
The answer according to the site for my input is 207, which given the
input is not the *fewest* number of steps to go from "e" to the medicine
molecule. Doing an exhaustive search is prohibitive, and I have not
invested in making the search more efficient, but there is at least one
solution in 206 steps, which can be seen, by setting the max
permutations to 1 million:

	$ ./day19 0 1 0 input
	207
	$ ./day19 0 1 1000000 input
	206

Presumably this is the result of a smaller key substitution allowing a
larger key substitution to follow.

Perhaps later I will log the steps, or construct fake data to
demonstrate this.

As the next-to-best answer is the answer accepted, I suspect I have
accomplished the intended goal of the task, which is to do an ordered
permutation search to find a *good* answer, if not the *best* answer.